+++ src/routes/leaderboard.tsx
@@ -7,18 +7,14 @@ import Card from '@/components/card'
 import { getUser } from '@/api/user/get-user'
 import { css } from 'linaria'
 import Authenticated from '@/components/authenticated'
+import { useQueryState } from '@/utils/use-query-state'
+import { useYears } from '@/utils/use-years'
+import { Dropdown } from '@/components/dropdown'
 
-const cardStyle = css`
-  padding: 0.6rem 1rem;
-  margin: 1rem;
-  margin-left: auto;
-  margin-right: auto;
-  max-width: 20rem;
-
-  & h1 {
-    font-weight: 500;
-    font-size: 1.2rem;
-  }
+const leaderboardCardTitleStyle = css`
+  font-weight: 500;
+  font-size: 1.2rem;
+  margin: 0.8rem;
 `
 
 const LeaderboardCard = ({
@@ -27,9 +23,9 @@ const LeaderboardCard = ({
   const user = usePromise(() => getUser(item.reporterId), [item])
 
   return (
-    <Card class={cardStyle}>
+    <Card>
       {user && (
-        <h1>
+        <h1 class={leaderboardCardTitleStyle}>
           {user.firstName} {user.lastName} - {item.reports}
         </h1>
       )}
@@ -37,12 +33,26 @@ const LeaderboardCard = ({
   )
 }
 
+const leaderboardListStyle = css`
+  display: grid;
+  grid-template-columns: 20rem;
+  justify-content: center;
+  grid-gap: 0.8rem;
+  padding: 0.8rem;
+`
+
+const currentYear = new Date().getFullYear()
 const LeaderboardList = () => {
-  const leaderboard = usePromise(getLeaderboard)
+  const [yearVal, setYear] = useQueryState('year', currentYear)
+  const year = Number(yearVal)
+  const years = useYears()
+  const leaderboard = usePromise(() => getLeaderboard(year), [year])
+
   if (!leaderboard) return <Spinner />
 
   return (
-    <div>
+    <div class={leaderboardListStyle}>
+      <Dropdown options={years} onChange={setYear} value={year} />
       {leaderboard.map(i => (
         <LeaderboardCard key={i.reporterId} item={i} />
       ))}
